requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
anthropic>=0.49.0
python-docx>=1.1.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
